requirements.txt
fastapi
transformers
torch
uvicorn
protobuf
sentencepiece
python-jose[cryptography]
python-multipart
